requirements.txt
requests
pdfplumber
neo4j
python-dotenv
tqdm
beautifulsoup4